# contracts/openapi/client-api.yaml
openapi: 3.1.0
info:
  title: DOM Client API
  version: 1.0.0
  description: >
    Official client HTTP contract for Account/Auth/Session and Workspace/Project/Folder
    metadata lifecycle (docs 09 and 28). This document describes only the HTTP surface:
    routes, operationId, headers and error-response assembly. Payload shapes are
    referenced from canonical JSON Schemas under /contracts and are never restated
    inline; code generation is driven by contracts/registry.yaml, not paths.
servers:
  - url: https://api.dom.internal
    description: Production
components:
  responses:
    ErrorEnvelopeResponse:
      description: Canonical error envelope (doc 28 §29)
      content:
        application/json:
          schema:
            $ref: "../errors/error-envelope.schema.json"
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Stable key for safe retry of a logical mutating command (doc 28 §13)
      schema:
        type: string
        format: uuid
paths:
...
  /v1/comments/{commentId}:
    patch:
      operationId: EditComment
      summary: Author edits their comment body (FR-CMT-003)
      tags: [comments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/comments/edit-comment.schema.json"
      responses:
        "200":
          description: Comment edited
          content:
            application/json:
              schema:
                $ref: "../commands/comments/edit-comment.schema.json#/$defs/EditCommentResponse"
        "400":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
    delete:
      operationId: DeleteComment
      summary: Author soft-deletes their comment (FR-CMT-004)
      tags: [comments]
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: "../commands/comments/delete-comment.schema.json#/$defs/DeleteCommentResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
  /v1/resources/{resourceId}/comments:
    post:
      operationId: AddComment
      summary: Append one flat-thread comment to a Resource (FR-CMT-001)
      tags: [comments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../commands/comments/add-comment.schema.json"
      responses:
        "201":
          description: Comment appended
          content:
            application/json:
              schema:
                $ref: "../commands/comments/add-comment.schema.json#/$defs/AddCommentResponse"
        "400":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
    get:
      operationId: ListComments
      summary: List flat-thread comments of a Resource (FR-CMT-002)
      tags: [comments]
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Comments oldest-first
          content:
            application/json:
              schema:
                $ref: "../queries/comments/list-comments.schema.json"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "403":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "503":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
...
  /v1/public/resources:
    get:
      operationId: GetPublicResources
      summary: Machine-client readable resources (PUB-002)
      tags: [public-api]
      parameters:
        - name: X-Dom-Key-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Dom-Signature
          in: header
          required: true
          schema:
            type: string
        - name: X-Dom-Timestamp
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                $ref: "../queries/public-api/list-public-resources.schema.json"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "429":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
  /v1/public/notifications:
    get:
      operationId: GetPublicNotifications
      summary: Machine-client notifications (integration-key auth, PUB-001)
      tags: [public-api]
      parameters:
        - name: X-Dom-Key-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Dom-Signature
          in: header
          required: true
          schema:
            type: string
        - name: X-Dom-Timestamp
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "../queries/public-api/list-public-notifications.schema.json"
        "401":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
        "429":
          $ref: "#/components/responses/ErrorEnvelopeResponse"
...